Add files via upload
We evaluate the proposed FTMSCL method on the text-dependent speaker recognition task by measuring a recognition accuracy. To do this, we first collect a smallscale FTM speech data set for training and evaluation. As sensitive piezoelectric film vibration sensor, LDT0-028K, manufactured by TE Technology, is used as the FTM for recording 
peech. The recording refers to the large-scale Google speech command data set which contains about 44,000 samples of 30 English words. We select the most common 20 words of the 30, and invite 20 speakers (10 males and 10 females) to pronounce them ten times, whose signals are recorded to form the FTM data set with 4,000 data. As a result, 4,000 FTM speech and CM speech samples are collected in a quiet environment, and the length of each sample is one second. After this, 80% of the data is divided for training and the rest for testing.