Update README.md
To do this, we first collect a small-scale FTM speech data set for training and evaluation.
As shown in Fig.1, a sensitive piezoelectric film vibration sensor, LDT0-028K, manufactured by TE Technology, is used as the FTM for recording speech. 
Google speech command data set that contains about 65,000 samples of 30 English words served as our reference in the process of recording FTM speech. 
We select the most common 20 words of the 30, and invite 20 speakers (10 males and 10 females) to pronounce them ten times.
The 20 speakers are from different parts of China and speak with different accents, which increases the diversity of this data set.
These signals are recorded to form the FTM data set with 4,000 data.